delete mock data file
